+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fixed issue where some C# modules would not run in Go agent
 -   Fixed SharpSploit/ShellCmd not running due to additional yaml argument
 -   Fixed install script failing on a subsequent run
+-   Fixed cookie naming for HTTP, foreign, and hop listeners
 
 ### Changed